requirements.txt
torch
torchvision
timm
albumentations
pandas
Pillow
scikit-learn
matplotlib
seaborn
tqdm
pyyaml
pytz
